Provide data store dependencies via Docker (#443)
Some specified versions were upgraded recently in CI builds:

* Postgres 12.3: https://github.com/rubyforgood/Flaredown/commit/f1175c324c2f9c4aa6d16f88997f9b00778f6886#diff-1fadaba7942eb03a370984f465dc59789417e580083e4bcaa2d073959b1815beR33
* MongoDB 4.4: https://github.com/rubyforgood/Flaredown/commit/f1175c324c2f9c4aa6d16f88997f9b00778f6886#diff-1fadaba7942eb03a370984f465dc59789417e580083e4bcaa2d073959b1815beR55

We opt here to match data store versions to prod. These are at most minor version upgrades past prod, and the target state for both staging and prod immediately.

We also correct a failure to connect to Mongo in local development which results from an inappropriate fallback when hunting for a cluster.