Allow users to find their account by email case-insensitive
One of the common support issues is that users try to reset their password, but get an error message because their account was not found. This is believed to often be caused by users typing their email address in with capital letters, even though all emails are stored in the database lower case.

This commit changes the user lookup to a case insensitive match, and adds test coverage for that.