Fix removing a reaction from a discussion post (#530)
* Allow DELETE requests in CORS

This fixes removing emoji from posts. It extends the whitelist of HTTP verbs. Future work should consider whether allowing all origins is a good idea

* Provide reaction IDs when available

A real reaction ID is only required/meaningful when the current user has reacted. Otherwise the value is meaningless. Use a random string to try to demonstrate that in cases where a reaction is present on the post, but the requested user has not reacted.

* Apply standardrb feedback