Add weather to existing export (#517)
* Add notes to existing export

* Add weather to data export

* remove notes from csv

* Fix staging URL typo

* Fix/no map reduce for reactions (#528)

* Add Reaction spec to cover current behavior

Getting ready to refactor `#values_count_with_participated`

* Refactor to avoid map_reduce

This implementation avoids map_reduce, which is not available on free/shared MongoDB tiers

* Reaction order is not important

Co-authored-by: Ben Langfeld <blangfeld@powerhrg.com>

* Buttons improvement summary page (#518)

* removed unnecessary background-color that is not used anymore on the buttons

* removed old background color variable and replaced with new one to match page styles

Co-authored-by: Ben Klang <bklang@wirehack.net>

* added background color and color filter buttons, matching styling throughout (#525)

* Fix removing a reaction from a discussion post (#530)

* Allow DELETE requests in CORS

This fixes removing emoji from posts. It extends the whitelist of HTTP verbs. Future work should consider whether allowing all origins is a good idea

* Provide reaction IDs when available

A real reaction ID is only required/meaningful when the current user has reacted. Otherwise the value is meaningless. Use a random string to try to demonstrate that in cases where a reaction is present on the post, but the requested user has not reacted.

* Apply standardrb feedback

* Update specs for new weather data

* Weather fields have to be included even if data is not present

Otherwise the column count in the data does not match the column count in the headers

* Add weather spec

Co-authored-by: Ben Langfeld <blangfeld@powerhrg.com>
Co-authored-by: Ben Klang <bklang@powerhrg.com>
Co-authored-by: Yunior Batista <yuniorbatista1113@gmail.com>
Co-authored-by: Ben Klang <bklang@wirehack.net>